requirements.txt
pillow==11.2.1
py-avataaars==1.1.2
streamlit==1.45.0
